+++ src/common/hooks/useLedger/useLedger.js
@@ -7,21 +7,31 @@ import {
 
 import { LEDGERS_API } from '../../const';
 
-export const useLedger = (ledgerId) => {
-  const ky = useOkapiKy();
+export const useLedger = (ledgerId, options = {}) => {
+  const {
+    enabled = true,
+    fiscalYearId,
+    tenantId,
+    ...queryParams
+  } = options;
+
+  const ky = useOkapiKy({ tenant: tenantId });
   const [namespace] = useNamespace({ key: 'ledger' });
 
   const {
     data = {},
     isFetching,
     isLoading,
-  } = useQuery(
-    [namespace, ledgerId],
-    async ({ signal }) => ky.get(`${LEDGERS_API}/${ledgerId}`, { signal }).json(),
-    {
-      enabled: Boolean(ledgerId),
+  } = useQuery({
+    queryKey: [namespace, ledgerId, fiscalYearId],
+    queryFn: ({ signal }) => {
+      const searchParams = fiscalYearId ? { fiscalYear: fiscalYearId } : undefined;
+
+      return ky.get(`${LEDGERS_API}/${ledgerId}`, { searchParams, signal }).json();
     },
-  );
+    enabled: enabled && Boolean(ledgerId),
+    ...queryParams,
+  });
 
   return ({
     ledger: data,